+++ internal/components/exit.go
@@ -1,0 +1,4 @@
+package components
+
+type Exit struct {
+}
