+++ internal/components/maze.go
@@ -1,0 +1,8 @@
+package components
+
+import "github.com/juanancid/maze-adventure/internal/layout"
+
+type Maze struct {
+	Layout   layout.Layout
+	CellSize int
+}
