+++ internal/components/position.go
@@ -1,0 +1,5 @@
+package components
+
+type Position struct {
+	X, Y float64
+}
