+++ internal/components/size.go
@@ -1,0 +1,5 @@
+package components
+
+type Size struct {
+	Width, Height float64
+}
