+++ internal/components/sprite.go
@@ -1,0 +1,7 @@
+package components
+
+import "github.com/hajimehoshi/ebiten/v2"
+
+type Sprite struct {
+	Image *ebiten.Image
+}
