+++ internal/components/velocity.go
@@ -1,0 +1,5 @@
+package components
+
+type Velocity struct {
+	DX, DY float64
+}
